Add coloration to the terrain
Contains a few extra architectural changes to try and make things extend
more nicely, as well. Looks great!

We set the texture's filterMode to Point in order to get rid of the
blurriness, and we set the wrapMode to Clamp so that we avoid any
wrapping-related discoloration of the sides.