Add height multiplier and mapping function
We make use of the very nice AnimationCurve that Lague uses in his
video. I abstract it a little away from the MeshGenerator though, so it
is just passed a Func<float, float> which maps each noise value to its
final vertex height. The function that MapGenerator passes to the
MeshGenerator uses the multiplier and AnimationCurve function.