Implement distance-based LOD
When the "viewer" is further away, a chunk will be rendered at a lower
LOD.

This change is pretty wild in terms how we accomplish this. Rather than
bundling the TerrainData (height map, colors) generation together with
the Mesh generation, we split that up into two different tasks that can
be put in their own TaskPools. This is useful as the same TerrainData
can be re-used to generate multiple LOD meshes - we don't want to repeat
that work for each LOD cause it'll be the exact same TerrainData.

To assist this, I implement a "LoadingCache". Initially I intended it to
operate as Guava's LoadingCache in Java, but I needed to add callbacks,
so it's pretty different. Basically, we give it a callback method when
loading something. If the value is already in the loading cache, we
callback immediately. Otherwise, we use *another* callback we were given
on construction for loading values. Once we receive our own callback
with the value, we then forward that to the callback we were handed in
the load method. Lots of callback-ing all around!

Rather than having a hard-coded max view distance, it's now determined
by the last element in the LodInfo array specified in the editor.

The last interesting thing is probably this ObjectPool class - another
fun creation I made! Despite AnimationCurve.Evaluate being annotated
with [ThreadSafeAttribute], it is not thread safe. I am noticing a ton
of artifacting on the generated mesh when I use it with multithreading.
When I single-thread, the artifacting disappears. If I multithread but
remove the animation curve, the artifacting is also not there.

So, what we do to address this is create a AnimationCurve ObjectPool.
Basically, we give each thread its own AnimationCurve instance. Each
will request one from the ObjectPool. If there's already one there, we
lend that to the thread, before putting it back in the pool. If there's
nothing in the pool, we create another AnimationCurve, use it, and put
it in the now-expanded pool.

This is a great way to allow us to use AnimationCurves in a
multi-threaded setting without instantiating lots of objects. Lague, in
his tutorial, just creates a new throw-away AnimationCurve for each use,
but this seems very wasteful, so this pool is quite an improvement!