Fix lighting seams between chunks
I spent like 2 evenings working on this... finally got it working! TL;DR
of the overall issue is that vertex nromals at the edges of chunks do
not take into account their neighbouring chunks, hence the lighting
would be off, causing visible seams between chunks.

This commit fixes that by "simulating" heights bordering each chunk and
thus calculating additional triangle normals that are not actually part
of the mesh. We then add these normals to the edge vertices and the
result is seamless lighting between chunks. Works at different LODs.

The issue that took me so long to realize was that, while we are
sampling the noise map from "bottom left", we actually want to
"simulate" the triangles from the top left. Hard to explain, but I
needed to therefore flips the Z/Y direction on the triangle edge vectors
to get the correct behavior. The issue manifested as the "simulated"
edge triangles having normals with a negated Z value of the correct one.