Generate collision mesh only when the player enters a chunk
A lot of lag spikes occur when moving around because we're loading so
many collision meshes as they spawn. This will reduce that amount (or
spread it out over time).

We also allow using a lower LOD for the collision mesh, further lowering
the lag spike, sacrificing an acceptable amount of collision accuracy.