Implement character jetpack
The player can now hold right click to boost up, as if they were using a
jetpack, getting us closer to the Tribes-style character controller. I
tried something different with this character controller - I think it
can be improved, but it works so far. The idea is to have the following
pipeline:

    PlayerInputController
    -> GroundAirMovementDelegator
        -> IF GROUNDED && !USING_JETPACK: Invoke GroundCharacterMover
        -> ELSE: Invoke AirCharacterMover

The Ground and Air character movers will assume they're being invoked
according to these rules.

I implement "PerFrameCache" as an additional generic utility so that we
only run the "is grounded" check once per frame. Another nice little
utility class, I think.

The next step is probably to implement the "spacebar to ski" control. I
can then tweak settings a bit and we should hopefully be quite close to
a quite good Tribes character controller! We will also need to implement
a limit to the fuel, a speed indicator, etc.

This commit also increases gravity from -9.81 to -20, to make the
character feel less floaty.