Add broken skiing with spacebar
The player can now "ski" with spacebar, but it is broken with this
commit. The idea is to get the normal to the terrain the player is on
and calculate a gravity-induced acceleration parallel to the terrain and
apply this to the player. I spent some time re-learning highschool
trigonometry and simplified things a little too much - I forgot about
the Z axis (so I'm currently only trying to handle X and Y
accelerations) and I also did not account for the fact that X on the
normal can legitimately be negative, so I don't think the math I'm doing
works out (I'm seeing some + y accelerations, which I would not expect).

You can sort of ski along the X world axis atm, but only in the negative
direction :P Will need to re-implement this, redoing the math.