Fix broken skiing mechanic
The previous commit added a knowingly-broken skiing mechanic, but it was
a start. We forgot about two elements that we fix in this commit,
namely:

1. We now take into account z axis.
2. We can calculate the acceleration correctly, even if components are
   negative.

We take advantage of Vector3.Cross in order to calculate a Vector3
direction down the slope. We then use some trigonometry I calculated on
paper to determine the magnitude to apply to that slope, and the result
is our acceleration vector. There's a brief proof (with some skipped
detail for ease of writing) in the code.

The skiing now works quite well! I will need to tweak the values a
little more and play around with things to get a more Tribes-like feel
that actually feels good to play with. I had to add a zero-friction
Physic Material to the player since I found that, while skiing, they
slow down too much. I put the friction to 0, but it still slows to a
stop, which I don't really understand. Will need to look into that.