Implement jetpack fuel
The jetpack has a limited fuel capacity that gets used up as the player
flies around. When the player runs out of fuel and stops using the
jetpack, it will automatically refill.

We currently only log the fuel tank's "fullness". I will add a UI
indicator in a later commit.