Add jetpack fuel bar to UI
Rather than debug logging the fuel tank's fullness, we now have a bar to
reflect that.

This involves introducing the idea of "GameEvents". I'm trying out a
singleton pattern (I know, I know) for a global events system, where any
part of the code can listen to and invoke any events. Of course you'd
only want e.g. the jetpack to invoke the "fuel changed" event, but
there's not much benefit to enforcing that in code right now.