Add updating text for player speed to UI
Add an additional little UI element to write the player's speed to the
screen. We treat each unit in the game as a meter, and using the
player's units per second, we calculate their "km per hour" speed.
Playing around with this, the values seem a little high, so I apply a
0.7 (atm) multiplier to this number before presenting the final "KPH" to
the player's screen.

Works well!