Add jetpack audio effect
As this is our first audio in the game, there's some setup to allow
defining sounds and playing them. I use a singleton AudioManager pattern
referenced via the GameManager, to allow playing sound from anywhere in
the scene. Notice that classes are referred to as "effects" for now. If
I want to do music, that should be handled separately later.