Update player's physic material to reflect skiing/walking state
Previously, the physic material was frictionless all the time (averaging
with the ground so there was *some* friction, to address the question
mark from a previous commit message). In this commit, we take the
minimum when combining the physic materials, meaning when the player
skiis, the friction is entirely decided by the player.

We add some scripting to that, when the player is not skiing, there's a
reasonable amount of friction and the player stops quickly. When the
player skiis, the friction drops dramatically (there's still a bit so
you can't slide forever on flat ground), allowing a nice skiing feel. We
also adjust the player's jetpack and movement values a bit in the scene
to get something closer to Tribes going. Still not entirely there, but
improving!